Make [jest-cli] not error when no tests are found with --findRelatedTests, --lastCommit or --onlyChanged options
