+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ## master
+
+### Features
+
+* `[jest-cli]` Make Jest exit without an error when no tests are found in
+  the case of `--lastCommit`, `--findRelatedTests`, or `--onlyChanged` options
+  having been passed to the CLI
+
 ## jest 22.0.6
 
 ### Fixes
